Update Visualisation
renaming visualization to turtle, table and graph
