+++ src/main/resources/templates/endpoint.mustache
@@ -75,9 +75,8 @@
 						<div class="col-sm-4">
 							<select class="form-control" name="format">
 								<option value="text/turtle">Turtle</option>
-								<option value="batik">Graphical 1</option>
-								<option value="d3">Graphical 2</option>
-								<option value="g6">Graphical 3</option>
+								<option value="table">Table</option>
+								<option value="graph">Graph</option>
 							</select>
 						</div>
 						<div class="col-sm-4">
