--- src/main/resources/templates/graphvisualisation_d3.mustache
@@ -1,35 +1,0 @@
-{{<super}}
-
-{{$title}}Revision Graph{{/title}}
-
-{{$head_extra}}
-    <link rel="stylesheet" href="/static/css/r43ples.versiongraph.css">
-    <link rel="stylesheet" href="/static/css/tipsy.css">
-{{/head_extra}}
- 
-{{$script_extra}}   
-<script src="/static/js/d3.min.js"></script>
-<script src="/static/js/dagre-d3.js"></script>
-<script src="/static/js/jquery.tipsy.js"></script>
-<script src="/static/js/version-graph.js"></script>
-
-<script type="text/javascript">
-    $(document).ready(function () {
-        drawGraph("#visualisation", "revisiongraph?graph={{graphName}}&format=application/json", false, false);
-    });
-</script>
-{{/script_extra}}
-    
-{{$content}}
-<div class="row">
-	<div class= "col-sm-12">
-	<h1>Revision Graph</h1>
-	    <h2><samp>&lt;{{graphName}}&gt;</samp></h2>
-		<div id="visualisation"></div>
-		<br>
-		<a class="btn btn-default" href="sparql">Back to Endpoint</a>
-	</div>
-</div>
-{{/content}}
-
-{{/super}}
