+++ src/main/resources/templates/graphvisualisation_graph.mustache
@@ -3,10 +3,9 @@
 {{$title}}Revision Graph{{/title}}
 
 {{$head_extra}}
-    <link rel="stylesheet" href="/static/css/r43ples.versiongraph2.css">
+    <link rel="stylesheet" href="/static/css/r43ples.versiongraph.css">
     <link rel="stylesheet" href="//code.jquery.com/ui/1.11.4/themes/smoothness/jquery-ui.css">
     <link rel="stylesheet" href="/static/css/jquery.qtip.css">
-    <link rel="stylesheet" href="/static/css/bootstrap.min.css">
 {{/head_extra}}
  
 {{$script_extra}}   
@@ -14,7 +13,7 @@
 <script src="/static/js/dagre-d3.js"></script>
 <script src="/static/js/jquery.qtip.js"></script>
 <script src="//code.jquery.com/ui/1.11.4/jquery-ui.js"></script>
-<script src="/static/js/version-graph2.js"></script>
+<script src="/static/js/version-graph.js"></script>
 <script src="/static/js/TurtleParser.js"></script>
 
 <script type="text/javascript">
