+++ src/main/resources/templates/graphvisualisation_table.mustache
