+++ src/main/resources/webapp/css/r43ples.versiongraph.css
@@ -1,7 +1,7 @@
 .revisionGraphVisualisation svg {
-	background-color: WhiteSmoke;
-	height: 468px;
-	border:1px solid #000000;
+	/*background-color: WhiteSmoke;*/
+	height: 368px;
+	border:1px solid #CBCBCB;
 	margin:1px;
 	overflow:hidden;
     cursor: default;
@@ -13,8 +13,53 @@
 }
 .revisionGraphVisualisation text {
     font-family: "Helvetica Neue", Helvetica, Arial, sans-serif;
-    font-size: 12px;
+    font-size: 14px;
 }
+ ul {
+  list-style: none;
+  margin-left: 0;
+  padding-left: 0;
+}
+li {
+  padding-left: 1em;
+  text-indent: -1em;
+}
+.properties th{
+	vertical-align: top;
+	padding-right: 10px;
+}
+#addsets, #deletesets {
+    width: 49.5%;
+    float:left;
+}
+#addsets{
+	margin-right: 1%; 
+} 
+.changeSetH{
+	width: 100%;
+    float:left;
+}
+.changeSetDiv{
+	width: 100%;
+    height: 200px;
+    overflow-y: scroll;
+	border: 1px solid #000000;
+}
+.changeSetTable{
+	width: 100%;
+	table-layout: fixed;
+}
+.changeSetTable td{
+	width: 33%;
+	word-wrap: break-word;
+	padding: 5px;
+}
+.changeSetTable tr:nth-child(odd) {
+	background-color: #f2f2f2
+}
+.changeSetTable tr:nth-child(odd) {
+	background-color: #f2f2f2
+} 
 
 .node circle {
     stroke: #333;
@@ -34,39 +79,15 @@
     pointer-events: none;
 }
 
-/* This styles the title of the tooltip */
-.tipsy h1, .tipsy h2 {
-    font-size: 1.6em;
-    font-weight: bold;
+.axis {
+  font: 10px sans-serif;
 }
 
-.tipsy h2 {
-    font-size: 1.4em;
+.x.axis  path, .x.axis line {
+  stroke: #000;
 }
-
-/* This styles the body of the tooltip */
-.tipsy-inner {
-	max-width: 800px;
-}
-
-.tipsy .properties {
-    font-size: 1.3em;
-    color: white;
-
-}
-
-
-.tipsy .properties td {
-	padding-left: 5px;
-}
-
-.tipsy .properties td:first-child {
-    font-weight: bold;
-    white-space: nowrap;
-    padding-left: 0px;
-}
-
-.tipsy a, .tipsy a:hover, .tipsy a:active {
-    color: #60b1fc;
-    text-decoration: none;
+.axis path,
+.axis line {
+  fill: none;
+  shape-rendering: crispEdges;
 }
