--- src/main/resources/webapp/css/r43ples.versiongraph2.css
@@ -1,93 +1,0 @@
-.revisionGraphVisualisation svg {
-	/*background-color: WhiteSmoke;*/
-	height: 368px;
-	border:1px solid #CBCBCB;
-	margin:1px;
-	overflow:hidden;
-    cursor: default;
-    -webkit-user-select: none;
-    -moz-user-select: none;
-    -ms-user-select: none;
-    -o-user-select: none;
-    user-select: none;
-}
-.revisionGraphVisualisation text {
-    font-family: "Helvetica Neue", Helvetica, Arial, sans-serif;
-    font-size: 14px;
-}
- ul {
-  list-style: none;
-  margin-left: 0;
-  padding-left: 0;
-}
-li {
-  padding-left: 1em;
-  text-indent: -1em;
-}
-.properties th{
-	vertical-align: top;
-	padding-right: 10px;
-}
-#addsets, #deletesets {
-    width: 49.5%;
-    float:left;
-}
-#addsets{
-	margin-right: 1%; 
-} 
-.changeSetH{
-	width: 100%;
-    float:left;
-}
-.changeSetDiv{
-	width: 100%;
-    height: 200px;
-    overflow-y: scroll;
-	border: 1px solid #000000;
-}
-.changeSetTable{
-	width: 100%;
-	table-layout: fixed;
-}
-.changeSetTable td{
-	width: 33%;
-	word-wrap: break-word;
-	padding: 5px;
-}
-.changeSetTable tr:nth-child(odd) {
-	background-color: #f2f2f2
-}
-.changeSetTable tr:nth-child(odd) {
-	background-color: #f2f2f2
-} 
-
-.node circle {
-    stroke: #333;
-    fill: #fff;
-    stroke-width:2px;
-}
-
-.node rect {
-    fill: #fff;
-}
-
-.edgePath path {
-    stroke-width: 2px;
-}
-
-.node text {
-    pointer-events: none;
-}
-
-.axis {
-  font: 10px sans-serif;
-}
-
-.x.axis  path, .x.axis line {
-  stroke: #000;
-}
-.axis path,
-.axis line {
-  fill: none;
-  shape-rendering: crispEdges;
-}
